+++ src/main/java/org/jsoup/safety/Cleaner.java
@@ -190,9 +190,17 @@ public class Cleaner {
         int numDiscarded = 0;
         Attributes sourceAttrs = sourceEl.attributes();
         for (Attribute sourceAttr : sourceAttrs) {
-            if (safelist.isSafeAttribute(sourceTag, sourceEl, sourceAttr))
-                destAttrs.put(sourceAttr);
-            else
+            if (safelist.isSafeAttribute(sourceTag, sourceEl, sourceAttr)) { // will keep this attr
+                String key = sourceAttr.getKey();
+                String value = sourceAttr.getValue();
+
+                if (safelist.shouldAbsUrl(sourceTag, key)) { // configured to make absolute urls for this key (href)
+                    value = sourceEl.absUrl(key);
+                    if (value.isEmpty()) // could not be made abs; leave as-is to allow custom unknown protocols
+                        value = sourceAttr.getValue();
+                }
+                destAttrs.put(key, value);
+            } else
                 numDiscarded++;
         }