+++ src/test/java/org/jsoup/safety/CleanerTest.java
@@ -7,6 +7,7 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Entities;
 import org.jsoup.nodes.Range;
+import org.jsoup.parser.ParseSettings;
 import org.jsoup.parser.Parser;
 import org.jsoup.parser.Tag;
 import org.jsoup.parser.TagSet;
@@ -282,6 +283,48 @@ public class CleanerTest {
         assertEquals("<a href=\"http://example.com/foo\">Link</a><img src=\"http://example.com/bar\">", clean);
     }
 
+    @Test void cleanDoesNotModifyInputDocumentWhenResolvingRelativeLinks() {
+        String html = "<a href='/foo'>Link</a>";
+        Cleaner cleaner = new Cleaner(Safelist.basic());
+        Document dirty = Jsoup.parseBodyFragment(html, "http://example.com/");
+        Document clean = cleaner.clean(dirty);
+
+        assertEquals("<a href=\"http://example.com/foo\">Link</a>", clean.body().html());
+        assertEquals("/foo", dirty.expectFirst("a").attr("href"));
+    }
+
+    @Test void isValidDoesNotModifyInputDocumentWhenResolvingRelativeLinks() {
+        String html = "<a href='/foo'>Link</a>";
+        Cleaner cleaner = new Cleaner(Safelist.basic());
+        Document dirty = Jsoup.parseBodyFragment(html, "http://example.com/");
+
+        assertTrue(cleaner.isValid(dirty));
+        assertEquals("<a href=\"http://example.com/foo\">Link</a>",
+            cleaner.clean(Jsoup.parseBodyFragment(html, "http://example.com/")).body().html());
+        assertEquals("/foo", dirty.expectFirst("a").attr("href"));
+    }
+
+    @Test void allPseudoTagProtocolsNormalizeRelativeLinks() {
+        Safelist safelist = new Safelist()
+            .addTags("a")
+            .addAttributes(":all", "href")
+            .addProtocols(":all", "href", "http", "https");
+        Document dirty = Jsoup.parseBodyFragment("<a href='/foo'>Link</a>", "http://example.com/");
+
+        assertEquals("<a href=\"http://example.com/foo\">Link</a>", new Cleaner(safelist).clean(dirty).body().html());
+    }
+
+    @Test void tagSpecificAttributesDoNotInheritAllPseudoTagProtocols() {
+        Safelist safelist = new Safelist()
+            .addTags("a")
+            .addAttributes("a", "href")
+            .addAttributes(":all", "href")
+            .addProtocols(":all", "href", "http", "https");
+        Document dirty = Jsoup.parseBodyFragment("<a href='/foo'>Link</a>", "http://example.com/");
+
+        assertEquals("<a href=\"/foo\">Link</a>", new Cleaner(safelist).clean(dirty).body().html());
+    }
+
     @Test public void preservesRelativeLinksIfConfigured() {
         String html = "<a href='/foo'>Link</a><img src='/bar'> <img src='javascript:alert()'>";
         String clean = Jsoup.clean(html, "http://example.com/", Safelist.basicWithImages().preserveRelativeLinks(true));
@@ -426,6 +469,18 @@ public class CleanerTest {
         Safelist relaxedWithAnchor = Safelist.relaxed().addProtocols("a", "href", "#");
         String clean = Jsoup.clean(dirty, relaxedWithAnchor);
         assertEquals("<a>One</a> <a>Two</a>", clean);
+    }
+
+    @Test void cleanerPreservesCaseVariantOfMatchingEnforcedAttribute() {
+        Document dirty = Jsoup.parse("<a href='http://example.com/' REL='nofollow'>Link</a>", "",
+            Parser.htmlParser().settings(ParseSettings.preserveCase));
+        Cleaner cleaner = new Cleaner(Safelist.basic());
+
+        Document clean = cleaner.clean(dirty);
+        Element link = clean.expectFirst("a");
+        assertTrue(link.hasAttr("REL"));
+        assertEquals("nofollow", link.attr("REL"));
+        assertTrue(cleaner.isValid(dirty));
     }
 
     @Test public void handlesNestedQuotesInAttribute() {